fix: global types file
