+++ .changeset/spicy-turtles-push.md
@@ -1,0 +1,5 @@
+---
+'polaris.shopify.com': patch
+---
+
+Fix global types file
